requirements.txt
alembic==1.13.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.12.0
Werkzeug==3.0.3
